tests/run-lisp-test.pl: make test fail if any output on stderr

+++ tests/run-lisp-test.pl
@@ -34,7 +34,7 @@ chomp $abs_srcdir;
 my $srcdir = $ENV{srcdir} || ".";
 my $builddir = $ENV{builddir} || ".";
 my $editor_name = $ENV{EDITOR_NAME};
-my @editor_cmd = split ' ', $ENV{EDITOR_CMD};
+my $editor_cmd = $ENV{EDITOR_CMD};
 
 # Compute test filenames
 my $test = shift;                       # Name is: ../tests/zile-only/backward_delete_char.el
@@ -48,9 +48,11 @@ $lisp_file =~ s/^\Q$srcdir/$abs_srcdir/e;
 
 # Perform the test
 mkpath(dirname($edit_file));
-push @editor_cmd, "--no-init-file", $edit_file, "--load", $lisp_file;
+my $cmd = "$editor_cmd --no-init-file \"$edit_file\" --load \"$lisp_file\"";
 copy("$srcdir/tests/test.input", $edit_file);
 chmod 0644, $edit_file;
-system(@editor_cmd) == 0 or die "$editor_name failed to run test `$name' with error code $?\n";
+my $stderr_output = `$cmd 3>&1 1>&2 2>&3 3>&-`; # Swap stderr & stdout https://www.perlmonks.org/?node_id=730
+die "$editor_name failed to run test `$name' with error code $?\n" unless $? == 0;
+die "$editor_name produced output on stderr:\n$stderr_output\n" if $stderr_output;
 system("diff", "$test.output", $edit_file) == 0 or die "$editor_name failed to produce correct output for test `$name'\n";
 unlink $edit_file, "$edit_file~";
